refactor(core): move api modules behind allowlist feature flags
