+++ .changes/api-feature-flags.md
@@ -1,0 +1,5 @@
+---
+"tauri": patch
+---
+
+Moves `shell`, `dialog::FileDialogBuilder` and `process::Command` APIs behind their allowlist feature flags.
