+++ core/tauri/Cargo.toml
@@ -33,12 +33,11 @@ normal = ["attohttpc"] # we ignore attohttpc because we can't remove it based on
 [dependencies]
 serde_json = { version = "1.0", features = [ "raw_value" ] }
 serde = { version = "1.0", features = [ "derive" ] }
-base64 = "0.13.0"
 tokio = { version = "1.5", features = [ "rt", "rt-multi-thread", "sync" ] }
 futures = "0.3"
-uuid = { version = "0.8.2", features = [ "v4" ] }
-thiserror = "1.0.24"
-once_cell = "1.7.2"
+uuid = { version = "0.8", features = [ "v4" ] }
+thiserror = "1.0"
+once_cell = "1.7"
 tauri-runtime = { version = "0.1.1", path = "../tauri-runtime" }
 tauri-macros = { version = "1.0.0-beta.1", path = "../tauri-macros" }
 tauri-utils = { version = "1.0.0-beta.0", path = "../tauri-utils" }
@@ -47,29 +46,44 @@ rand = "0.8"
 tempfile = "3"
 semver = "0.11"
 serde_repr = "0.1"
-dirs-next = "2.0.0"
-zip = "0.5.12"
-ignore = "^0.4.17"
-either = "1.6.1"
+zip = "0.5"
+ignore = "0.4"
+either = "1.6"
 tar = "0.4"
 flate2 = "1.0"
-rfd = "0.3.0"
 tinyfiledialogs = "3.3"
 http = "0.2"
-clap = { version = "=3.0.0-beta.2", optional = true }
-notify-rust = { version = "4.5.0", optional = true }
-tauri-hotkey = { version = "0.1.2", optional = true }
-open = "1.7.0"
-shared_child = "0.3"
-os_pipe = "0.9"
-minisign-verify = "0.1.8"
 state = "0.4"
 bincode = "1.3"
+dirs-next = "2.0"
+
+# FS
+base64 = { version = "0.13", optional = true } # also used on the updater
+
+# CLI
+clap = { version = "=3.0.0-beta.2", optional = true }
+
+# Notifications
+notify-rust = { version = "4.5.0", optional = true }
+
+# Global shortcut
+tauri-hotkey = { version = "0.1.2", optional = true }
 
 # HTTP
 reqwest = { version = "0.11", features = [ "json", "multipart" ], optional = true }
 bytes = { version = "1", features = [ "serde" ], optional = true }
 attohttpc = { version = "0.17", features = [ "json", "form" ] }
+
+# Shell
+open = { version = "1.7", optional = true }
+shared_child = { version = "0.3", optional = true }
+os_pipe = { version = "0.9", optional = true }
+
+# Dialogs
+rfd = { version = "0.3", optional = true }
+
+# Updater
+minisign-verify = { version = "0.1", optional = true }
 
 [build-dependencies]
 cfg_aliases = "0.1.1"
@@ -90,31 +104,31 @@ dox = [ "tauri-runtime-wry/dox" ]
 wry = [ "tauri-runtime-wry" ]
 cli = [ "clap" ]
 custom-protocol = [ "tauri-macros/custom-protocol" ]
-api-all = [ "notification-all", "global-shortcut-all", "updater" ]
-updater = [ ]
+api-all = [ "notification-all", "global-shortcut-all", "shell-all", "dialog-all", "updater" ]
+updater = [ "minisign-verify", "base64" ]
 menu = [ "tauri-runtime/menu", "tauri-runtime-wry/menu" ]
-system-tray = [ "tauri-runtime/system-tray", "tauri-runtime-wry/system-tray" ]
+system-tray = [ "tauri-runtime/system-tray", "tauri-runtime-wry/system-tray"]
 reqwest-client = [ "reqwest", "bytes" ]
 fs-all = [ ]
 fs-read-text-file = [ ]
 fs-read-binary-file = [ ]
 fs-write-file = [ ]
-fs-write-binary-file = [ ]
+fs-write-binary-file = [ "base64" ]
 fs-read-dir = [ ]
 fs-copy-file = [ ]
 fs-create-dir = [ ]
 fs-remove-dir = [ ]
 fs-remove-file = [ ]
 fs-rename-file = [ ]
-fs-path-api = [ ]
+fs-path = [ ]
 window-all = [ ]
 window-create = [ ]
-shell-all = [ ]
-shell-execute = [ ]
-shell-open = [ ]
-dialog-all = [ ]
-dialog-open = [ ]
-dialog-save = [ ]
+shell-all = [ "shell-open", "shell-execute" ]
+shell-execute = [ "shared_child", "os_pipe" ]
+shell-open = [ "open" ]
+dialog-all = [ "dialog-open", "dialog-save" ]
+dialog-open = [ "rfd" ]
+dialog-save = [ "rfd" ]
 http-all = [ ]
 http-request = [ ]
 notification-all = [ "notify-rust" ]
