+++ examples/helloworld/src-tauri/Cargo.toml
@@ -10,7 +10,7 @@ tauri-build = { path = "../../../core/tauri-build", features = [ "codegen" ] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = [ "derive" ] }
-tauri = { path = "../../../core/tauri", features = ["api-all"] }
+tauri = { path = "../../../core/tauri", features = [] }
 
 [features]
 default = [ "custom-protocol" ]
