+++ examples/helloworld/src-tauri/tauri.conf.json
@@ -35,7 +35,7 @@
       }
     },
     "allowlist": {
-      "all": true
+      "all": false
     },
     "windows": [
       {
